% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBLW.R
\name{blw.mean.descendants}
\alias{blw.mean.descendants}
\title{Mean distance from internal nodes to descendant tips}
\usage{
blw.mean.descendants(tree)
}
\arguments{
\item{tree}{a \code{phylo} class tree object that is binary (see \code{\link[ape]{multi2di}})}
}
\description{
Calculates the mean distance from each internal node to its descendant tips
}
\details{
This is a function used by \code{\link{calculate.blw}} when \code{method='mean.descendants'}.
}

