% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedILR.R
\name{buildilrBasep}
\alias{buildilrBasep}
\title{Weighted ILR Contrast Matrix}
\usage{
buildilrBasep(W, p)
}
\arguments{
\item{W}{sequantial binary partition matrix (e.g., signary matrix; output of
\link{phylo2sbp})}

\item{p}{weights (should not be closed)}
}
\value{
matrix
}
\description{
Weighted ILR Contrast Matrix
}
\author{
Justin Silverman (adapted from \link[compositions]{gsi.buildilrBase})
}
\references{
J. J. Egozcue, V. Pawlowsky-Glahn. \emph{Changing the Reference
  Measure in the Simplex and its Weighting Effects}. Austrian Journal of
  Statistics, 2016
}

