% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBLW.R
\name{calculate.blw}
\alias{calc.blw}
\alias{calculate.blw}
\title{Calculate Branch Length Weightings for ILR Coordinates}
\usage{
calculate.blw(tree, method = "sum.children", tip.boosted = NULL)
}
\arguments{
\item{tree}{a \code{phylo} class tree object that is binary (see \code{\link[ape]{multi2di}})}

\item{method}{options include: (default) \code{'sum.children'} and \code{'mean.descendants'}
see details for more information.}

\item{tip.boosted}{(Optional) numeric factor to multiply tip lengths by to 'boost their importance
in the weightings. \emph{Note:} this feature is largely unstudied.}
}
\value{
vector of weightings for ILR coordinates produced via specified method.
}
\description{
Calculates the weightings for ILR coordinates based on branch lenghts of
a phylogenetic tree via a few different methods (see details).
}
\details{
ILR balances built from a binary partition of a phylogenetic tree
can be imbued with branch length information. This function is helpful in
calculating those weightings.\cr \cr
There are a number of methods for calculating these weightings, the default
\code{'sum.children'} calculates the weighting for a given balance as the sum
of its two direct children's branch length. An alternative that has been as yet less
studied is \code{'mean.descendants'} to calculate the weighting for a given balance
as the sum of its two direct children's branch lengths PLUS for each child the average
distance from it to its descendant tips. \cr \cr
\emph{Note:} That some trees contain tips with branch lengths of zero length. This can result
in that tip being unreasonably downweighted, as such this function automatically
adds a small pseudocount to those tips with zero length (equal to the smallest non-zero)
branch length on the tree.
}
\examples{
library(phyloseq)
data(CSS)
tree <- phy_tree(CSS)
calculate.blw(tree, method='sum.children')[1:10]
calculate.blw(tree, method='mean.descendants')[1:10]
}
\author{
Justin Silverman
}
\seealso{
\code{\link{philr}}
}

