% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_ILR.R
\name{clrp}
\alias{clrp}
\title{Weighted CLR Transform}
\usage{
clrp(y, p)
}
\arguments{
\item{y}{shifted data matrix (e.g., output of \link{shiftp})}

\item{p}{weights (should not be closed)}
}
\value{
matrix
}
\description{
Weighted CLR Transform
}
\details{
Note that this function will close the dataset \code{y} to 1.
}
\examples{
p <- seq(.1,1,by=.2)
c <- t(rmultinom(10,100,c(.1,.6,.2,.3,.2))) + 0.65   # add a small pseudocount
x <- miniclo(c)
y <- shiftp(x, p)
clrp(y, p)
}
\author{
Justin Silverman
}
\references{
J. J. Egozcue, V. Pawlowsky-Glahn (2016) \emph{Changing the Reference
  Measure in the Simplex and its Weighting Effects}. Austrian Journal of
  Statistics 45(4):25-44
}

