% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CSS}
\alias{CSS}
\title{Caporaso Skin Sites dataset}
\format{A \code{phyloseq} object containing OTU table and paired, tree, taxonomy,
and metadata. The key sample variable is \code{'BODY_SITE'}}
\source{
Personal Communication with Dan Knights
}
\usage{
CSS
}
\description{
This is the Costello Skin Sites dataset (OTU table, tree, taxonomy, and metadata)
used in Knights, D., 2011 paper on benchmarks for supervised classification of human microbiota.
}
\references{
Costello EK, Lauber CL, Hamady M, Fierer N, Jeffrey I, Gordon JI & Knight R (2009)
Bacterial community variation in human body habitats across space and time. Science 326: 1694–1697.

Knights, D., Costello, E. K. & Knight, R. Supervised classification of human
microbiota. FEMS Microbiol. Rev. 35, 343–359 (2011).
}
\keyword{datasets}

