% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_ILR.R
\name{g.colMeans}
\alias{g.colMeans}
\title{Geometric Means of Columns}
\usage{
g.colMeans(x)
}
\arguments{
\item{x}{matrix or vector}
}
\value{
vector (geometric mean of columns)
}
\description{
Calculates geometric mean of columns. Does not calculate
WEIGHTED geometric means (vs. \link{g.rowMeans})
}
\examples{
philr:::g.colMeans(rbind(c(2,4,4), c(2,4,4)))
}
\seealso{
g.rowMeans
}
