% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedILR.R
\name{ilrp}
\alias{ilrp}
\title{Weighted ILR Transform}
\usage{
ilrp(y, p, V)
}
\arguments{
\item{y}{shifted data matrix (e.g., output of \link{shiftp})}

\item{p}{weights (should not be closed)}

\item{V}{weighted contrast matrix (e.g., output of \link{buildilrBasep})}
}
\value{
matrix
}
\description{
Calculated using weighted CLR transform (\link{clrp})
}
\author{
Justin Silverman
}
\references{
J. J. Egozcue, V. Pawlowsky-Glahn. \emph{Changing the Reference
  Measure in the Simplex and its Weighting Effects}. Austrian Journal of
  Statistics, 2016
}

