% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_ILR.R
\name{miniclo}
\alias{miniclo}
\title{miniclo}
\usage{
miniclo(c, k = 1)
}
\arguments{
\item{c}{dataset to be closed}

\item{k}{closure constant}
}
\value{
matrix (if c is a vector or matrix) or data.frame (if c is a data.frame)
}
\description{
small function to close (aka normalize by proportions, aka total sum scaling)
a dataset to a constant \code{k} (usually taken to be 1). After closure the row sums
of the dataset should sum to \code{k}.
}
\examples{
c <- matrix(c(1,2,3,1,2,3,1,2,3), nrow = 3, byrow=TRUE)
miniclo(c)
miniclo(c, k=2)
}

