% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{named_rtree}
\alias{named_rtree}
\title{Generate random tree with named internal nodes}
\usage{
named_rtree(n)
}
\arguments{
\item{n}{an integer giving the number of tips in the tree.}
}
\value{
An object of class "phylo"
}
\description{
Internal nodes are named by numbering and adding the prefix 'n'. This
function is laregly for use in examples throughout this package.
}

