% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo2sbp.R
\name{phylo2sbp}
\alias{phylo2sbp}
\title{Create Sequential Binary Partition from Phylogenetic Tree}
\usage{
phylo2sbp(tr, n_cores = 1)
}
\arguments{
\item{tr}{a \code{phylo} tree object with n leaves}

\item{n_cores}{(Optional) integer specifying the number of cores to use. See Details.}
}
\value{
a n by n-1 matrix of the sequential binary partition sign matrix
}
\description{
This function converts a binary phylogenetic tree to sequential binary
parition to be used to then build an ILR basis for compositional
metagenomic data.
}
\details{
Parallelization is done through \code{parallel} package using type "FORK".
Note parallelization is rarely needed, even for trees of upwards of 40,000 leaves.
}
\examples{
tr <- named_rtree(5)
phylo2sbp(tr)
}
\author{
Justin Silverman
}
\seealso{
\code{\link{philr}}
}

