% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_balance}
\alias{plot_balance}
\title{Plot a Balance}
\usage{
plot_balance(coord.name, tr, tax = NULL, plot.tax = NULL,
  color.tax = NULL, ...)
}
\arguments{
\item{coord.name}{the name of a balance/internal node on the tree (given as a string)}

\item{tr}{(Optional) though needed if \code{name.balance=TRUE}}

\item{tax}{a matrix/data.frame of taxonomy, rownames should correspond to \code{tr$tip.labels}
columns should be taxonomic levels (named) with increasing taxonomic resolution from left to right
(e.g., Phylum to the left of Genus).}

\item{plot.tax}{(Optional) Plot bar/heatmap of taxonomy labels at specified level
(e.g., column name in \code{tax})}

\item{color.tax}{taxonomic level (e.g., column name in \code{tax}) to color the tree by}

\item{...}{pass other arguments to ggtree (e.g., \code{layout='fan'})}

\item{color.tax}{(Optional) passed to gheatmap \code{color}}
}
\value{
plot created with ggtree
}
\description{
Visualize a balance in place on a phylogenetic tree (numerator of balance is in green, denominator in blue)
}
\details{
\code{tax} only needs to be specified if \code{color.tax != NULL}
}
\examples{
data(CSS)
df <- CSS$otu.table + 0.65   # add a small pseudocount
tree <- CSS$phy.tree
df.philr <- philr(df, tree, part.weights='anorm.x.gm.counts',
                  ilr.weights='blw.sqrt', return.all=FALSE, n_cores=1)
df.philr.long <- convert_to_long(df.philr, CSS$sample.data$BODY_SITE)
plot_balance('n7', tree)
plot_balance('n7', tree, layout='fan')
plot_balance('n7', tree, tax=CSS$tax.table, color.tax='Phylum')

# plot with phyla bar/strip
plot_balance('n7', tree, CSS$tax.table, plot.tax=c('Phylum'))
}

