% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_density_breakdown_wtree}
\alias{plot_density_breakdown_wtree}
\title{Plot Balance on Tree next to Density Breakdown}
\usage{
plot_density_breakdown_wtree(df, coord.name, tax, tr, name.balance = TRUE,
  plot.tax = NULL, color.tax = NULL, ...)
}
\arguments{
\item{df}{PhILR transformed data (see \code{\link{philr}}) in long format
(see \code{\link{convert_to_long}})}

\item{coord.name}{the name of a balance/internal node on the tree (given as a string)}

\item{tax}{a matrix/data.frame of taxonomy, rownames should correspond to \code{tr$tip.labels}
columns should be taxonomic levels (named) with increasing taxonomic resolution from left to right
(e.g., Phylum to the left of Genus).}

\item{tr}{(Optional) though needed if \code{name.balance=TRUE}}

\item{name.balance}{logical (default: \code{FALSE}) of whether title should include output from
call to \code{\link{name.balance}}}

\item{plot.tax}{(Optional) Plot bar/heatmap of taxonomy labels at specified level
(e.g., column name in \code{tax})}

\item{color.tax}{taxonomic level (e.g., column name in \code{tax}) to color the tree by}

\item{...}{pass other arguments to ggtree (e.g., \code{layout='fan'})}
}
\value{
nothing returned, plot called with \code{multiplot}
}
\description{
Visualizes Density breakdown \code{\link{plot_density_breakdown}} and balance in place on a phylogenetic tree
\code{\link{plot_balance}} (numerator of balance is in green, denominator in blue) next to each other.
}
\examples{
data(CSS)
df <- t(phyloseq::otu_table(CSS))
df <- df + 0.65   # add a small pseudocount
tree <- phyloseq::phy_tree(CSS)
df.philr <- philr(df, tree, part.weights='anorm.x.gm.counts',
                  ilr.weights='blw.sqrt', return.all=FALSE, n_cores=1)
df.philr.long <- convert_to_long(df.philr, get_variable(CSS, 'BODY_SITE'))
plot_density_breakdown_wtree(df.philr.long, 'n7', tax_table(CSS), tree)
plot_density_breakdown_wtree(df.philr.long, 'n7', tax_table(CSS), tree, plot.tax='Phylum')
}

