% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedILR.R
\name{clrp}
\alias{clrp}
\title{Weighted CLR Transform}
\usage{
clrp(y, p)
}
\arguments{
\item{y}{shifted data matrix (e.g., output of \link{shiftp})}

\item{p}{weights (should not be closed)}
}
\value{
matrix
}
\description{
Weighted CLR Transform
}
\author{
Justin Silverman
}
\references{
J. J. Egozcue, V. Pawlowsky-Glahn. \emph{Changing the Reference
  Measure in the Simplex and its Weighting Effects}. Austrian Journal of
  Statistics, 2016
}

