% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{convert_to_long}
\alias{convert_to_long}
\title{Converts wide format ILR transformed data to long format}
\usage{
convert_to_long(df, labels)
}
\arguments{
\item{df}{PhILR transformed data in wide format (samples by balances) (see \code{\link{philr}})}

\item{labels}{vector (of length \code{nrow(df)}) with labels to group samples by}
}
\value{
\code{df} in long format with columns
\itemize{
\item sample
\item labels
\item coord
\item value
}
}
\description{
Converts wide format ILR transformed data (see \code{\link{philr}}) to long format
useful in various plotting functions
(e.g., \code{\link{plot_density_breakdown_wtree}})
}
\examples{
data(CSS)
df <- t(phyloseq::otu_table(CSS))
df <- df + 0.65   # add a small pseudocount
tree <- phyloseq::phy_tree(CSS)
df.philr <- philr(df, tree, part.weights='anorm.x.gm.counts',
                  ilr.weights='blw.sqrt', return.all=FALSE, n_cores=1)
head(convert_to_long(df.philr, get_variable(CSS, 'BODY_SITE'))
}

