% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namebalance.R
\name{name.balance}
\alias{name.balance}
\title{Name a balance (coordinate) based on taxonomy}
\usage{
name.balance(tr, tax, coord, method = c("voting"), thresh = 0.95)
}
\arguments{
\item{tr}{an object of class \code{'phylo'}}

\item{tax}{a matrix/data.frame of taxonomy, rownames should correspond to \code{tr$tip.labels}
columns should be taxonomic levels (named) with increasing taxonomic resolution from left to right
(e.g., Phylum to the left of Genus).}

\item{coord}{the name of a balance/internal node on the tree (given as a string)}

\item{method}{currently only \code{'voting'} implemented. See Details.}

\item{thresh}{threshold for assignment of taxonomy to a given part of a balance (see details).}
}
\description{
For a given ILR balance (coordinate) assigns a name to the balance based on a
provided taxonomy table. This is useful for interpretation of the balances.
}
\details{
The method \code{'voting'} assigns the name of the each part of a balance (e.g., numerator and denominator / each
child of \code{coord}) as follows:
\enumerate{
\item First Subset \code{tax} to contain only descendent tips of the given child of \code{coord}
\item Second At the finest taxonomic (farthest right of \code{tax}) see if any one taxonomic label
is present at or above \code{thresh}. If yes output that taxonomic label (at that taxonomic level) as the label
for that child of \code{coord}. If no then move to coarser taxonomic level (leftward) and repeat.
}
}
\seealso{
\code{\link{philr}}
}

