% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_balance}
\alias{plot_balance}
\title{Plot a Balance}
\usage{
plot_balance(coord.name, tr, p.ggtree = NULL, ...)
}
\arguments{
\item{coord.name}{the name of a balance/internal node on the tree (given as a string)}

\item{tr}{(Optional) though needed if \code{name.balance=TRUE}}

\item{p.ggtree}{a \code{ggtree} object, for very large trees (>10,000) consider
precomputing the basic tree as part of the plot. Usually not necessesary}

\item{...}{pass other arguments to ggtree (e.g., \code{layout='fan'})}
}
\value{
plot created with ggtree
}
\description{
Visualize a balance in place on a phylogenetic tree (numerator of balance is in green, denominator in blue)
}
\examples{
library(phyloseq)
data(CSS)
df <- t(otu_table(CSS))
df <- df + 0.65   # add a small pseudocount
tree <- phy_tree(CSS)
df.philr <- philr(df, tree, part.weights='anorm.x.gm.counts',
                  ilr.weights='blw.sqrt', return.all=FALSE, n_cores=1)
df.philr.long <- convert_to_long(df.philr, get_variable(CSS, 'BODY_SITE'))
plot_balance('n7', tree)
plot_balance('n7', tree, layout='fan')
}

